refactor(tracking): rename infrastructure in tracking module
