+++ entrypoint/src/main/scala/io.github.positionpal.location.entrypoint/Launcher.scala
@@ -10,9 +10,9 @@ import io.github.positionpal.location.application.groups.impl.BasicUserGroupsSer
 import io.github.positionpal.location.application.sessions.impl.BasicUsersSessionService
 import io.github.positionpal.location.commons.ConfigurationError
 import io.github.positionpal.location.grpc.{GrpcServer, GrpcUserSessionService}
-import io.github.positionpal.location.infrastructure.services.ActorBasedRealTimeTracking
-import io.github.positionpal.location.infrastructure.services.projections.UserSessionProjection
-import io.github.positionpal.location.infrastructure.utils.AkkaUtils
+import io.github.positionpal.location.tracking.services.ActorBasedRealTimeTracking
+import io.github.positionpal.location.tracking.services.projections.UserSessionProjection
+import io.github.positionpal.location.tracking.utils.AkkaUtils
 import io.github.positionpal.location.messages.groups.RabbitMQGroupsEventConsumer
 import io.github.positionpal.location.messages.{MessageBrokerConnectionFactory, RabbitMQ}
 import io.github.positionpal.location.presentation.proto.UserSessionServiceFs2Grpc
