--- infrastructure/src/main/scala/io/github/positionpal/location/infrastructure/services/actors/AkkaSerializable.scala
@@ -1,7 +1,0 @@
-package io.github.positionpal.location.infrastructure.services.actors
-
-/** Marker trait for remote serializable messages exchanged between distributed actors.
-  * Every message class exchanged between distributed actors or persisted should extend this trait.
-  * This trait should be referenced as binding in the Akka serialization configuration.
-  */
-trait AkkaSerializable
