+++ tracking-actors/src/main/resources/db-scripts/create-tables.cql
