+++ tracking-actors/src/test/resources/testable-akka-config.conf
@@ -1,0 +1,18 @@
+# =================================================================================================
+# A minimal Akka configuration for testing purposes (see `actors` package)
+# this is not complete and should not be used in production
+# =================================================================================================
+akka.actor.provider = "cluster"
+akka.remote.artery.canonical {
+  hostname = "127.0.0.1"
+  port = 0
+}
+akka.actor {
+  serializers {
+    borer-json = "io.github.positionpal.location.tracking.actors.BorerAkkaSerializer"
+  }
+  serialization-bindings {
+    "io.github.positionpal.location.tracking.actors.AkkaSerializable" = borer-json
+    "io.github.positionpal.location.domain.DrivingEvent" = borer-json
+  }
+}
